+++ 003/src/my_func.c
@@ -1,7 +1,7 @@
 #include "head.h"
 
 // функция вычисления факториала
-long int factorial (long int n) {
+unsigned long long int factorial (unsigned long long int n) {
   if ((n == 0) || (n == 1)) { // проверяем на равенство 0 или 1
     return 1; // если равно выходим
   }
